prepare for byte instruction, change instruction signature to uint16

+++ pdpcpu/cpu.go
@@ -68,12 +68,12 @@ type CPU struct {
 	// the opcode function should append to the following signature:
 	// param: instruction int16
 	// return: error -> nil if everything went OK
-	singleOpOpcodes       map[uint16](func(int16) error)
-	doubleOpOpcodes       map[uint16](func(int16) error)
-	rddOpOpcodes          map[uint16](func(int16) error)
-	controlOpcodes        map[uint16](func(int16) error)
-	singleRegisterOpcodes map[uint16](func(int16) error)
-	otherOpcodes          map[uint16](func(int16) error)
+	singleOpOpcodes       map[uint16](func(uint16) error)
+	doubleOpOpcodes       map[uint16](func(uint16) error)
+	rddOpOpcodes          map[uint16](func(uint16) error)
+	controlOpcodes        map[uint16](func(uint16) error)
+	singleRegisterOpcodes map[uint16](func(uint16) error)
+	otherOpcodes          map[uint16](func(uint16) error)
 }
 
 /**
@@ -102,12 +102,12 @@ func New(mmunit *mmu.MMU18Bit) *CPU {
 	c.doubleTrap = false
 
 	// single operand
-	c.singleOpOpcodes = make(map[uint16](func(int16) error))
-	c.doubleOpOpcodes = make(map[uint16](func(int16) error))
-	c.rddOpOpcodes = make(map[uint16](func(int16) error))
-	c.controlOpcodes = make(map[uint16](func(int16) error))
-	c.otherOpcodes = make(map[uint16](func(int16) error))
-	c.singleRegisterOpcodes = make(map[uint16](func(int16) error))
+	c.singleOpOpcodes = make(map[uint16](func(uint16) error))
+	c.doubleOpOpcodes = make(map[uint16](func(uint16) error))
+	c.rddOpOpcodes = make(map[uint16](func(uint16) error))
+	c.controlOpcodes = make(map[uint16](func(uint16) error))
+	c.otherOpcodes = make(map[uint16](func(uint16) error))
+	c.singleRegisterOpcodes = make(map[uint16](func(uint16) error))
 
 	// single opearnd:
 	c.singleOpOpcodes[0100] = c.jmpOp
@@ -206,7 +206,7 @@ func (c *CPU) Fetch() uint16 {
 // if instruction matching the mask not found in the opcodes map, fallback and try
 // to match anything lower.
 // Fail ultimately.
-func (c *CPU) Decode(instr uint16) func(int16) error {
+func (c *CPU) Decode(instr uint16) func(uint16) error {
 	// 2 operand instructions:
 	var opcode uint16
 
@@ -256,7 +256,7 @@ func (c *CPU) Decode(instr uint16) func(int16) error {
 func (c *CPU) Execute() error {
 	instruction := c.Fetch()
 	opcode := c.Decode(instruction)
-	return opcode(int16(instruction))
+	return opcode(instruction)
 }
 
 // helper functions:
